# docs/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Music Library API",
        "description": "API для онлайн библиотеки песен.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local server"
        }
    ],
    "paths": {
        "/songs": {
            "get": {
                "summary": "Получить список песен",
                "description": "Получение данных библиотеки с возможностью фильтрации по полям и пагинацией.",
                "tags": ["songs"],
                "parameters": [
                    {
                        "name": "group",
                        "in": "query",
                        "description": "Фильтрация по названию группы",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "song",
                        "in": "query",
                        "description": "Фильтрация по названию песни",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "release_date",
                        "in": "query",
                        "description": "Фильтрация по дате выпуска",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Количество элементов на странице",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список песен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Song"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при получении списка песен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Добавить новую песню",
                "description": "Добавление новой песни в базу данных.",
                "tags": ["songs"],
                "requestBody": {
                    "description": "Данные новой песни",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Song"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Добавленная песня",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Song"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при добавлении песни",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/songs/{id}": {
            "get": {
                "summary": "Получить песню по ID",
                "description": "Извлечение песни по ее уникальному идентификатору.",
                "tags": ["songs"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о песне",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Song"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Песня не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Обновить песню",
                "description": "Обновление существующей песни по ее ID.",
                "tags": ["songs"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Обновленные данные песни",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Song"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Обновленная песня",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Song"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Песня не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Удалить песню",
                "description": "Удаление песни по ее ID.",
                "tags": ["songs"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Песня успешно удалена"
                    },
                    "404": {
                        "description": "Песня не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при удалении песни",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Song": {
                "type": "object",
                "required": ["group", "song"],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "group": {
                        "type": "string",
                        "example": "Muse"
                    },
                    "song": {
                        "type": "string",
                        "example": "Supermassive Black Hole"
                    },
                    "releaseDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2006-07-16"
                    },
                    "text": {
                        "type": "string",
                        "example": "Ooh baby, don't you know I suffer?\nOoh baby, can you hear me moan?"
                    },
                    "link": {
                        "type": "string",
                        "example": "https://www.youtube.com/watch?v=Xsp3_a-PMTw"
                    }
                }
            },
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Ошибка сервера"
                    }
                }
            }
        }
    }
}